# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Cartlabs API
  version: 0.2.0
  description: Public REST contract for Cartlabs web and API applications.
servers:
  - url: /api/v1
...
security:
  - BearerAuth: []
paths:
...
  /admin/stores/{storeId}/moderation:
    patch:
      operationId: moderateStore
      summary: Moderate store
      tags: [Admin]
      parameters:
        - {$ref: "#/components/parameters/StoreId"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ModerationInput"}
      responses:
        "200":
          description: Moderated store
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Store"}
        "404": {$ref: "#/components/responses/NotFound"}
...
components:
  securitySchemes:
    BearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
    RefreshCookie:
      type: apiKey
      in: cookie
      name: __Host-cartlabs_refresh
      description: Production Secure cookie with Path=/. Local HTTP uses cartlabs_refresh scoped to authentication routes.
  parameters:
...
    StoreId: {name: storeId, in: path, required: true, schema: {type: string, format: uuid}}
...
  responses:
...
    NotFound: {$ref: "#/components/responses/Problem404"}
...
    Problem404: {description: Resource not found, content: {application/problem+json: {schema: {$ref: "#/components/schemas/Problem"}}}}
...
  schemas:
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status]
      properties:
        type: {type: string, format: uri-reference}
        title: {type: string}
        status: {type: integer, minimum: 400, maximum: 599}
        detail: {type: string}
        requestId: {type: string}
...
    ModerationStatus: {type: string, enum: [pending, approved, rejected]}
...
    StoreInput:
      type: object
      additionalProperties: false
      required: [name, slug, description]
      properties:
        name: {type: string, minLength: 2, maxLength: 100}
        slug: {type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'}
        description: {type: string, maxLength: 1000}
    Store:
      allOf:
        - {$ref: "#/components/schemas/StoreInput"}
        - type: object
          required: [id, sellerId, status, moderationNote, createdAt, updatedAt]
          properties:
            id: {type: string, format: uuid}
            sellerId: {type: string, format: uuid}
            status: {$ref: "#/components/schemas/ModerationStatus"}
            moderationNote: {type: string}
            createdAt: {type: string, format: date-time}
            updatedAt: {type: string, format: date-time}
...
    ModerationInput:
      type: object
      additionalProperties: false
      required: [status, note]
      properties:
        status: {type: string, enum: [approved, rejected]}
        note: {type: string, maxLength: 500}